refactor(entry): lib/types/{entry,calendar-event} を features/entry/types/ へ移動
cross-layer に置かれていた Entry 関連型の owner を確定し、
src/lib/types を曖昧な型置き場から外す。

- lib/types/entry.ts (EntryState / EntryOrigin / FulfillmentScore) を
  features/entry/types/entry.ts に inline
- lib/types/calendar-event.ts (CalendarEvent) を
  features/entry/types/calendar-event.ts へ移動 (Entry の表示射影型のため
  owner は entry。calendar → entry は既存の cross-feature 方向)
- features/entry barrel に 4 type を追加
- features/calendar/types/calendar.types.ts は CalendarEvent を
  features/entry barrel 経由で re-export
- lib/types/settings.ts は shell / lib stores から参照される genuine
  cross-layer のため移動せず維持
- 型 only の移動、runtime / 仕様 / UI / DB 変更なし
