+++ apps/product/src/features/entry/types/calendar-event.ts
